Fix token values in mock states
