+++ tests/constants/settings.js
@@ -1,11 +1,5 @@
 const settings = {
   areTermsOfUseAccepted: true,
-  token: {
-    list: {
-      LSK: true,
-      BTC: false,
-    },
-  },
 };
 
 export default settings;
